rename intersection
