+++ index.ts
@@ -331,7 +331,7 @@ import { WebGLFilterBackend } from './src/filters/webgl_backend.class';
 import { runningAnimations } from './src/util/animation/AnimationRegistry';
 import { Observable } from './src/Observable';
 import { Point } from './src/point.class';
-import { Intersection } from './src/intersection.class';
+import { Intersection } from './src/Intersection';
 import { Color } from './src/color/Color';
 import { Control } from './src/controls/Control';
 import { Gradient } from './src/gradient/Gradient';
