+++ src/Intersection.ts
