+++ src/shapes/Object/ObjectGeometry.ts
@@ -7,7 +7,7 @@ import type {
   TOriginY,
 } from '../../typedefs';
 import { iMatrix } from '../../constants';
-import { Intersection } from '../../intersection.class';
+import { Intersection } from '../../Intersection';
 import { Point } from '../../point.class';
 import { makeBoundingBoxFromPoints } from '../../util/misc/boundingBoxFromPoints';
 import { cos } from '../../util/misc/cos';
